pyproject.toml
[project]
name = "agent-proj"
version = "0.2.0"
description = "Multi-agent travel planning system with Phoenix observability, Nemo deployment, and comprehensive evaluation (FREE APIs only)"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arize>=7.51.2",
    "langchain>=1.2.3",
    "langchain-core>=1.2.7",
    "langchain-openai>=1.1.7",
    "langgraph>=1.0.6",
    "pydantic>=2.10.*",
    "python-dotenv>=1.2.1",
    "reportlab>=4.4.7",
    "requests>=2.31.0",
    "arize-phoenix>=12.30.0",
    "streamlit>=1.53.0",
]
